refactor(suite-native): move AccountLabel to accounts package
