+++ suite-native/accounts/src/components/__tests__/AccountLabel.test.tsx
