+++ suite-native/labeling/package.json
@@ -39,8 +39,5 @@
         "react": "19.2.4",
         "react-native": "0.83.2",
         "react-redux": "9.2.0"
-    },
-    "devDependencies": {
-        "@suite-native/test-utils-store": "workspace:*"
     }
 }
