+++ suite-native/labeling/src/index.ts
@@ -4,8 +4,7 @@ export { SendFormLabelEditable } from './components/SendFormLabelEditable';
 export { TransactionOutputLabel } from './components/TransactionOutputLabel';
 export { TransactionOutputLabelEditable } from './components/TransactionOutputLabelEditable';
 export { WalletLabel } from './components/WalletLabel';
-export { selectIsLabellingAllowed, selectAccountLabel } from './selectors';
 export type { CombinedLabelingState } from './selectors';
-export { AccountLabel } from './components/AccountLabel';
+export { selectIsLabellingAllowed } from './selectors';
 export { useTurnOnSuiteSyncGuard } from './hooks/useTurnOnSuiteSyncGuard';
 export { useSuiteSyncErrorHandler } from './hooks/useSuiteSyncLabelErrorHandler';
