+++ suite-native/module-earn/package.json
@@ -46,7 +46,6 @@
         "@suite-native/helpers": "workspace:*",
         "@suite-native/icons": "workspace:*",
         "@suite-native/intl": "workspace:*",
-        "@suite-native/labeling": "workspace:*",
         "@suite-native/link": "workspace:*",
         "@suite-native/message-system": "workspace:*",
         "@suite-native/navigation": "workspace:*",
