+++ suite-native/module-trading/src/components/general/MyAssetSheet/MyAssetListSectionHeader.tsx
@@ -2,10 +2,13 @@ import { useSelector } from 'react-redux';
 
 import { type AccountsRootState, selectFormattedAccountType } from '@suite-common/wallet-core';
 import { type Account } from '@suite-common/wallet-types';
-import { type NativeAccountsRootState, selectAccountFiatBalance } from '@suite-native/accounts';
+import {
+    AccountLabel,
+    type NativeAccountsRootState,
+    selectAccountFiatBalance,
+} from '@suite-native/accounts';
 import { Badge, HStack, Text } from '@suite-native/atoms';
 import { BaseCurrencyAmountFormatter } from '@suite-native/formatters';
-import { AccountLabel } from '@suite-native/labeling';
 
 export type MyAssetListSectionHeaderProps = {
     account: Account;
