+++ suite-native/receive/src/components/ReceiveScreenHeader.tsx
@@ -4,9 +4,9 @@ import { getNetworkDisplaySymbol } from '@suite-common/wallet-config';
 import { type AccountsRootState, selectAccountNetworkSymbol } from '@suite-common/wallet-core';
 import { type AccountKey, type TokenAddress } from '@suite-common/wallet-types';
 import { parseAccountKey } from '@suite-common/wallet-utils';
+import { AccountLabel } from '@suite-native/accounts';
 import { HStack, Text } from '@suite-native/atoms';
 import { Translation } from '@suite-native/intl';
-import { AccountLabel } from '@suite-native/labeling';
 import {
     type CloseActionType,
     ScreenHeader,
