+++ suite-native/receive/tsconfig.json
@@ -29,7 +29,6 @@
         { "path": "../icons" },
         { "path": "../in-app-rating" },
         { "path": "../intl" },
-        { "path": "../labeling" },
         { "path": "../link" },
         { "path": "../module-device-onboarding" },
         { "path": "../navigation" },
